+++ broker/lib/client_error.py
@@ -1,0 +1,6 @@
+class ClientError(Exception):
+    """
+    This class is used for errors that have messaging that clients are allowed to see.
+    """
+
+    pass
